clean cli package
