+++ packages/cli/src/fetchProjectMetadata.ts
@@ -1,9 +1,4 @@
-import pc from "picocolors";
-import {
-  projectDetailSchema,
-  type ProjectDetail,
-  type ProjectFile,
-} from "@transi-store/common";
+import { projectDetailSchema, type ProjectDetail } from "@transi-store/common";
 
 // Extract the useful part of a network-level fetch error. `fetch failed` on
 // its own hides the actual reason (connection refused, DNS error, TLS) —
@@ -22,17 +17,19 @@ export function describeFetchError(error: unknown): string {
   return error.message;
 }
 
+type FetchProjectMetadataOptions = {
+  domainRoot: string;
+  apiKey: string;
+  org: string;
+  project: string;
+};
+
 export async function fetchProjectMetadata({
   domainRoot,
   apiKey,
   org,
   project,
-}: {
-  domainRoot: string;
-  apiKey: string;
-  org: string;
-  project: string;
-}): Promise<ProjectDetail> {
+}: FetchProjectMetadataOptions): Promise<ProjectDetail> {
   const url = `${domainRoot}/api/orgs/${org}/projects/${project}`;
 
   let response: Response;
@@ -63,49 +60,4 @@ export async function fetchProjectMetadata({
     );
   }
   return parsed.data;
-}
-
-// Replace the <lang> placeholder in a file path with an actual locale.
-export function resolveFilePath(filePath: string, locale: string): string {
-  return filePath.replace("<lang>", locale);
-}
-
-// Pick a file from the project's metadata. When fileIdArg is omitted and the
-// project has exactly one file, that file is returned. Otherwise the CLI must
-// exit with a helpful message, so this function calls process.exit on error.
-export function pickFile(
-  files: Array<ProjectFile>,
-  fileIdArg: string | undefined,
-  projectName: string,
-): ProjectFile {
-  if (files.length === 0) {
-    console.error(pc.red(`Project "${projectName}" has no files configured`));
-    process.exit(1);
-  }
-
-  const firstFile = files[0]!;
-  if (fileIdArg === undefined) {
-    if (files.length === 1) {
-      return firstFile;
-    }
-    console.error(
-      pc.red(
-        `Project "${projectName}" has ${files.length} files — use --file <id> to pick one.`,
-      ),
-    );
-    for (const f of files) {
-      console.error(pc.dim(`  ${f.id}\t${f.filePath}\t(${f.format})`));
-    }
-    process.exit(1);
-  }
-
-  const id = Number.parseInt(fileIdArg, 10);
-  const found = files.find((f) => f.id === id);
-  if (!found) {
-    console.error(
-      pc.red(`File id "${fileIdArg}" not found in project "${projectName}".`),
-    );
-    process.exit(1);
-  }
-  return found;
 }
